# services/hotel/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hotel service for microservices.",
        "title": "Swagger Hotel API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/{country_code}/{city_slug}/hotels/": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get hotels from admin or moderator provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotels"
                ],
                "summary": "Get hotels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Code",
                        "name": "country_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "City Slug",
                        "name": "city_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "default": 20,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.HotelList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new hotel from admin provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotels"
                ],
                "summary": "Create hotel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Code",
                        "name": "country_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "City Slug",
                        "name": "city_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hotel data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_request.HotelCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.Hotel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/{country_code}/{city_slug}/hotels/{hotel_slug}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get hotel by slug from admin, moderator or owner provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotels"
                ],
                "summary": "Get hotel by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Code",
                        "name": "country_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "City Slug",
                        "name": "city_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hotel slug",
                        "name": "hotel_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.Hotel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update hotel by slug from admin, moderator or owner provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotels"
                ],
                "summary": "Update hotel by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Code",
                        "name": "country_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "City Slug",
                        "name": "city_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hotel slug",
                        "name": "hotel_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hotel data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_request.HotelUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.HotelUpdate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete hotel by slug from admin or owner provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotels"
                ],
                "summary": "Delete hotel by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Code",
                        "name": "country_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "City Slug",
                        "name": "city_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hotel slug",
                        "name": "hotel_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    }
                }
            }
        },
...
        "/{country_code}/{city_slug}/hotels/{hotel_slug}/update_title": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update hotel title by slug from admin, moderator or owner provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotels"
                ],
                "summary": "Update hotel title by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Code",
                        "name": "country_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "City Slug",
                        "name": "city_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hotel slug",
                        "name": "hotel_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hotel data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_request.HotelTitleUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.HotelTitleUpdate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/hotel_internal_http_dto_response.ErrorSchema"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helper.PaginationLinks": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                }
            }
        },
        "hotel_internal_http_dto_request.HotelCreate": {
            "type": "object",
            "required": [
                "address",
                "location",
                "owner_id",
                "title"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 300,
                    "minLength": 5
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "location": {
                    "$ref": "#/definitions/hotel_internal_http_dto_request.Location"
                },
                "owner_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
        "hotel_internal_http_dto_request.HotelTitleUpdate": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
        "hotel_internal_http_dto_request.HotelUpdate": {
            "type": "object",
            "required": [
                "address",
                "location"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 300,
                    "minLength": 5
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "location": {
                    "$ref": "#/definitions/hotel_internal_http_dto_request.Location"
                }
            }
        },
        "hotel_internal_http_dto_request.Location": {
            "type": "object",
            "required": [
                "latitude",
                "longitude"
            ],
            "properties": {
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180
                }
            }
        },
...
        "hotel_internal_http_dto_response.ErrorSchema": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "hotel_internal_http_dto_response.Hotel": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/hotel_internal_http_dto_response.Location"
                },
                "owner_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "hotel_internal_http_dto_response.HotelList": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "hotels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/hotel_internal_http_dto_response.HotelShort"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "links": {
                    "$ref": "#/definitions/helper.PaginationLinks"
                },
                "total_page_count": {
                    "type": "integer"
                },
                "total_rooms_count": {
                    "type": "integer"
                }
            }
        },
        "hotel_internal_http_dto_response.HotelShort": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/hotel_internal_http_dto_response.Location"
                },
                "owner_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "hotel_internal_http_dto_response.HotelTitleUpdate": {
            "type": "object",
            "properties": {
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "hotel_internal_http_dto_response.HotelUpdate": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/hotel_internal_http_dto_response.Location"
                }
            }
        },
        "hotel_internal_http_dto_response.Location": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}